STR-1343 commit aggregated sig to state and serve via rpc (#143)
* feat(duty-tracker): commit aggregated sigs

* feat(rpc): add rpc impl to serve signature for challenge/disprove

* refactor(rpc): use typed error codes

* refactor(executor): use committed agg sig in PayoutOptimistic

* fix(stake-chain): replace IndexSet with BTreeMap as insertion order is unreliable

* refactor(csm): move nonce aggregation and partials verification to separate fn

* tests(csm): add test for nonce aggregation

* fix(csm): use the right graph input and cache key

* fix(csm): use cached graph

* chore(tx-driver): log confirmation

* docs: fix typos and add cross refs

* docs(rpc): add docs to RPC traits

* refactor(duty-tracker): rename ouroboros_event* to synthetic_event*

* fix(duty-tracker): improve log message

* refactor(duty-tracker): use separate enum for synthetic events

* refactor(duty-tracker): mux synthetic event stream into contract events

* detach synthetic event handling from csm

---------

+++ crates/duty-tracker/src/errors.rs
@@ -84,7 +84,7 @@ pub enum StakeChainErr {
 
     /// Error indicating that some stake chain data is missing.
     #[error("stake chain inputs incomplete for operator: {0}, index: {1}")]
-    IncompleteStakeChainInput(P2POperatorPubKey, usize),
+    IncompleteStakeChainInput(P2POperatorPubKey, u32),
 
     /// Error indicating that the pre stake data for the operator does not exist in the database.
     #[error("stake setup data not found for operator: {0}")]
